+++ src/core/f-qsort.c
@@ -194,7 +194,8 @@ loop:	SWAPINIT(a, es);
 	pn = (char *)a + n * es;
 	r = min(pa - (char *)a, pb - pa);
 	vecswap(a, pb - r, r);
-	r = min(pd - pc, pn - pd - es);
+	// !!! Ren/C: pn - pd - es => (long)(pn - pd - ps) for -Wsign-compare
+	r = min(pd - pc, (long)(pn - pd - es));
 	vecswap(pb, pn - r, r);
 	if ((r = pb - pa) > es)
 #ifdef I_AM_QSORT_R
